pyproject.toml
[tool.poetry]
name = "Swifty Daily Scrum Bot"
version = "0.1.0"
description = "write daily scrum"
authors = ["msh4gr@gmail.com"]

[tool.poetry.dependencies]
python = "^3.9"
discord-py = "^2.3.2"
tzdata = ">=2023.4"
flask = "^3.0.1"

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"